[LYW] 01/22 문제풀이
